SPORTS-33 Renamed folders in frontend/src
